Moving the add-hook inside the enable function.
It is a convention to try to avoid modifying the user environment without
their consent. By evaluating `(auth-source-kwallet-enable)` the user is giving
some kind of content, and thus the hook should be added there (or so I think!).

Commenting why using the `auth-source-backend-parser-functions` hook instead of
advising the `auth-source-backend-parse` function.